+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/likes/dto/CommentLikeRequest.kt
@@ -1,0 +1,5 @@
+package com.teamsparta.exhibitionnewsfeed.domain.likes.dto
+
+data class CommentLikeRequest(
+    val userId: Long
+)
