+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/likes/service/LikeService.kt
@@ -5,4 +5,7 @@ import com.teamsparta.exhibitionnewsfeed.domain.likes.dto.PostLikeRequest
 interface LikeService {
     fun likePost(postId: Long, request: PostLikeRequest)
     fun removePostLike(likeId: Long, postId: Long)
+
+    fun likeComment(postId: Long, commentId: Long, userId: Long)
+    fun removeCommentLike(postId: Long, commentId: Long, userId: Long, likeId: Long)
 }
