+++ src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/likes/service/LikeServiceImpl.kt
@@ -1,8 +1,11 @@
 package com.teamsparta.exhibitionnewsfeed.domain.likes.service
 
 import com.teamsparta.exhibitionnewsfeed.domain.likes.dto.PostLikeRequest
+import com.teamsparta.exhibitionnewsfeed.domain.likes.model.CommentLike
 import com.teamsparta.exhibitionnewsfeed.domain.likes.model.PostLike
+import com.teamsparta.exhibitionnewsfeed.domain.likes.repository.CommentLikeRepository
 import com.teamsparta.exhibitionnewsfeed.domain.likes.repository.PostLikeRepository
+import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.repository.CommentRepository
 import com.teamsparta.exhibitionnewsfeed.domain.newsfeed.post.repository.PostRepository
 import com.teamsparta.exhibitionnewsfeed.domain.user.repository.UserRepository
 import com.teamsparta.exhibitionnewsfeed.exception.ModelNotFoundException
@@ -12,6 +15,8 @@ import org.springframework.stereotype.Service
 
 @Service
 class LikeServiceImpl(
+    private val commentLikeRepository: CommentLikeRepository,
+    private val commentRepository: CommentRepository,
     private val postLikeRepository: PostLikeRepository,
     private val postRepository: PostRepository,
     private val userRepository: UserRepository
@@ -43,5 +48,41 @@ class LikeServiceImpl(
 
         val like = postLikeRepository.findByIdOrNull(likeId) ?: throw ModelNotFoundException("postLike", likeId)
         postLikeRepository.delete(like)
+    }
+
+    @Transactional
+    override fun likeComment(postId: Long, commentId: Long, userId: Long) {
+        validatePost(postId)
+
+        val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        val user = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User", userId)
+        val isExistLike = commentLikeRepository.existsByCommentAndUser(comment, user)
+        if (isExistLike) {
+            throw IllegalStateException("Cannot add a like when it has already")
+        }
+
+        commentLikeRepository.save(
+            CommentLike(
+                comment = comment,
+                user = user
+            )
+        )
+    }
+
+    @Transactional
+    override fun removeCommentLike(postId: Long, commentId: Long, userId: Long, likeId: Long) {
+        validatePost(postId)
+
+        val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
+        val user = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User", userId)
+        val isExistLike = commentLikeRepository.existsByCommentAndUser(comment, user)
+        if (isExistLike.not()) {
+            throw IllegalStateException("Cannot remove a like when it has already")
+        }
+        commentLikeRepository.deleteById(likeId)
+    }
+
+    private fun validatePost(postId: Long) {
+        postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
     }
 }
