--- src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/newsfeed/comment/dto/CommentLikeRequest.kt
@@ -1,5 +1,0 @@
-package com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.dto
-
-data class CommentLikeRequest(
-    val userId: Long
-)
