--- src/main/kotlin/com/teamsparta/exhibitionnewsfeed/domain/newsfeed/comment/service/CommentLikeService.kt
@@ -1,7 +1,0 @@
-package com.teamsparta.exhibitionnewsfeed.domain.newsfeed.comment.service
-
-interface CommentLikeService {
-    fun addLike(postId: Long, commentId: Long, userId: Long)
-
-    fun removeLike(postId: Long, commentId: Long, userId: Long, likeId: Long)
-}
